Add homepage hack for relative urls
To get temporary builds that are not located at the root s3
directory (eg s3://osma-health/long-sha-here/index.html) we need
to have relative routing of the assets.

Create react app assumes that the app is hosted at the "homepage"
key of package.json, defaulting to "/". If you specify "." as
the homepage, then it tricks CRA into using relative paths.

!! This only works if client-side routing is handled using a hash
based router. PushState routers will not work with this method.

If we switch to a PushState router, we need to remove the homepage
app and figure out a mechanism to deploy at the root.